Issue 59: List item's text can't be started from *.
Function s:process_tag_list() searches the line for *bold* text at the start of
line. If there is *bold* text then skip the line -- it is not list item (which
could be started by a star).

The problem was in bold text regexp. Two stars with spaces between (* *) were
treated as bold whereas it is not.

I have fixed bold, italic, bolditalic and italicbold regexps to not include a
single space between corresponding symbols.