Fix issue 295.  Discovered and fixed *MAJOR* issue with VimwikiAll2HTML. ...
Before this revision, VimwikiAll2HTML did not load each file before converting
to HTML.  As a result, the chain of events:

  s:tag_wikilink() > vimwiki#base#resolve_scheme() > vimwiki#base#current_subdir()

resulted in the incorrect subdirectory being returned (... for testing, this
was typically empty).

This resolves the contradictory results that I was getting when loading
individual test files and comparing results.